+++ packages/api/core/src/api/publish.ts
@@ -165,7 +165,7 @@ const publish = async ({
       await asyncOra(`Resolving publish target: ${`${resolvablePublishTarget.name}`.cyan}`, async () => { // eslint-disable-line no-loop-func
         PublisherClass = requireSearch(dir, [resolvablePublishTarget.name]);
         if (!PublisherClass) {
-          throw new Error(`Could not find a publish target with the name: ${resolvablePublishTarget.name}`);
+          throw new Error(`Could not find a publish target with the name: ${resolvablePublishTarget.name}. Make sure it's listed in the devDependencies of your package.json`);
         }
       });