+++ packages/api/core/test/slow/api_spec_slow.ts
@@ -197,7 +197,7 @@ describe(`electron-forge API (with installer=${nodeInstaller})`, () => {
 
       const packageJSON = await readRawPackageJson(dir);
       packageJSON.name = 'testapp';
-      packageJSON.productName = 'Test App';
+      packageJSON.productName = 'Test-App';
       packageJSON.config.forge.packagerConfig.asar = false;
       if (process.platform === 'win32') {
         await fs.copy(
@@ -264,14 +264,14 @@ describe(`electron-forge API (with installer=${nodeInstaller})`, () => {
     });
 
     describe('after package', () => {
-      let resourcesPath = 'Test App.app/Contents/Resources';
+      let resourcesPath = 'Test-App.app/Contents/Resources';
       if (process.platform !== 'darwin') {
         resourcesPath = 'resources';
       }
 
       it('should have deleted the forge config from the packaged app', async () => {
         const cleanPackageJSON = JSON.parse(asar.extractFile(
-          path.resolve(dir, 'out', `Test App-${process.platform}-${process.arch}`, resourcesPath, 'app.asar'),
+          path.resolve(dir, 'out', `Test-App-${process.platform}-${process.arch}`, resourcesPath, 'app.asar'),
           'package.json',
         ));
         expect(cleanPackageJSON).to.not.have.nested.property('config.forge');