+++ packages/maker/flatpak/test/MakerFlatpak_spec.ts
@@ -19,7 +19,7 @@ class MakerImpl extends MakerBase<MakerFlatpakConfig> {
 describe('MakerFlatpak', () => {
   let MakerFlatpak: typeof MakerImpl;
   let maker: MakerImpl;
-  let eidStub: SinonStub;
+  let eifStub: SinonStub;
   let ensureDirectoryStub: SinonStub;
   let config: MakerFlatpakConfig;
   let createMaker: () => void;
@@ -32,12 +32,12 @@ describe('MakerFlatpak', () => {
 
   beforeEach(() => {
     ensureDirectoryStub = stub().returns(Promise.resolve());
-    eidStub = stub().callsArg(1);
+    eifStub = stub().resolves();
     config = {};
 
     MakerFlatpak = proxyquire.noPreserveCache().noCallThru().load('../src/MakerFlatpak', {
       'fs-extra': { readdir: stub().returns(Promise.resolve([])) },
-      'electron-installer-flatpak': eidStub,
+      '@malept/electron-installer-flatpak': eifStub,
     }).default;
     createMaker = () => {
       maker = new MakerFlatpak(config);
@@ -49,9 +49,9 @@ describe('MakerFlatpak', () => {
 
   it('should pass through correct defaults', async () => {
     await (maker.make as any)({
-      dir, makeDir, appName, targetArch, packageJSON,
+      dir, makeDir, appName, targetArch, packageJSON
     });
-    const opts = eidStub.firstCall.args[0];
+    const opts = eifStub.firstCall.args[0];
     expect(opts).to.deep.equal({
       arch: flatpakArch(process.arch as ForgeArch),
       src: dir,
@@ -71,7 +71,7 @@ describe('MakerFlatpak', () => {
     await (maker.make as any)({
       dir, makeDir, appName, targetArch, packageJSON,
     });
-    const opts = eidStub.firstCall.args[0];
+    const opts = eifStub.firstCall.args[0];
     expect(opts).to.deep.equal({
       arch: flatpakArch(process.arch as ForgeArch),
       options: {