+++ packages/publisher/base/src/Publisher.ts
@@ -1,5 +1,5 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
-import { ForgePlatform, ForgeConfig, ForgeMakeResult } from '@electron-forge/shared-types';
+import { ForgePlatform, ForgeConfig, ForgeMakeResult, IForgePublisher } from '@electron-forge/shared-types';
 
 export interface PublisherOptions {
   /**
@@ -18,7 +18,7 @@ export interface PublisherOptions {
   forgeConfig: ForgeConfig;
 }
 
-export default abstract class Publisher<C> {
+export default abstract class Publisher<C> implements IForgePublisher {
   public abstract name: string;
 
   public defaultPlatforms?: ForgePlatform[];
@@ -37,7 +37,7 @@ export default abstract class Publisher<C> {
   get platforms() {
     if (this.providedPlatforms) return this.providedPlatforms;
     if (this.defaultPlatforms) return this.defaultPlatforms;
-    return ['win32', 'linux', 'darwin', 'mas'];
+    return ['win32', 'linux', 'darwin', 'mas'] as ForgePlatform[];
   }
 
   /**